requirements.txt
alembic==1.13.1
blinker==1.7.0
Bootstrap-Flask==2.3.3
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
pillow==10.2.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
tzdata==2024.1
Werkzeug==3.0.1
WTForms==3.1.2
